LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-ordering-expected.txt


FAIL Multiple async-fetched CSS modules maintain attribute order in adoptedStyleSheets. assert_equals: Before fetch: expected 3 placeholder(s). expected 3 but got 0
